II. Java. Часть 2. Classes

+++ app/src/main/java/com/kozyrev/simbirtraineeship/classes/ClassesTask3/Triangle.java
@@ -2,11 +2,11 @@ package com.kozyrev.simbirtraineeship.classes.ClassesTask3;
 
 public class Triangle {
 
-    private int x1, x2, x3, y1, y2, y3;
+    private double x1, x2, x3, y1, y2, y3;
     private double firstSide, secondSide, thirdSide;
-    private int firstAngle, secondAngle, thirdAngle;
+    private double firstAngle, secondAngle, thirdAngle;
 
-    public Triangle(int x1, int y1, int x2, int y2, int x3, int y3, int firstAngle, int secondAngle, int thirdAngle) throws Exception{
+    public Triangle(double x1, double y1, double x2, double y2, double x3, double y3, double firstAngle, double secondAngle, double thirdAngle) throws Exception{
         this.x1 = x1;
         this.x2 = x2;
         this.x3 = x3;
@@ -24,7 +24,7 @@ public class Triangle {
         thirdSide = this.side(x3, x1, y3, y1);
     }
 
-    public double side(int x1, int x2, int y1, int y2){
+    public double side(double x1, double x2, double y1, double y2){
         return Math.pow((Math.pow((x2 - x1), 2)+ Math.pow((x2 - x1), 2)), 0.5);
     }
 
@@ -41,12 +41,12 @@ public class Triangle {
         return new double[]{(x1 + x2 + x3) / 3, (y1 + y2 + y3) / 3};
     }
 
-    public int[] getPoints(){
-        return new int[]{x1, y1, x2, y2, x3, y3};
+    public double[] getPoints(){
+        return new double[]{x1, y1, x2, y2, x3, y3};
     }
 
-    public int[] getAngles(){
-        return new int[]{firstAngle, secondAngle, thirdAngle};
+    public double[] getAngles(){
+        return new double[]{firstAngle, secondAngle, thirdAngle};
     }
 
     public double[] getSides(){
